requirements.txt
asgiref==3.7.2
Django==5.1-rc1
djangorestframework==3.15.2
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
